feat: реализовать централизованную обработку ошибок

+++ controllers/cards.js
@@ -1,84 +1,51 @@
 const Card = require('../models/card');
 
-const {
-  INCORRECT_DATA,
-  NOT_FOUND,
-  sendServerErrorMessage,
-  sendInvalidIdMessage,
-  sendCardNotFoundByIdMessage,
-} = require('../utils/utils');
+const IncorrectDataError = require('../errors/incorrect-data-err');
+const ForbiddenError = require('../errors/forbidden-err');
+const NotFoundError = require('../errors/not-found-err');
 
-module.exports.getCards = (req, res) => {
+module.exports.getCards = (req, res, next) => {
   Card.find({})
     .then((cards) => res.send({ data: cards }))
-    .catch(() => sendServerErrorMessage(res));
+    .catch((err) => next(err));
 };
 
-module.exports.createCard = (req, res) => {
+module.exports.createCard = (req, res, next) => {
   const { user: { _id }, body: { name, link } } = req;
 
   Card.create({ name, link, owner: _id })
     .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(INCORRECT_DATA).send({ message: 'Ошибка! Переданы некорректные данные при создании карточки' });
-        return;
+        next(new IncorrectDataError('Ошибка! Переданы некорректные данные при создании карточки'));
       }
-      sendServerErrorMessage(res);
+      next(err);
     });
 };
 
-module.exports.deleteCard = (req, res) => {
+module.exports.deleteCard = (req, res, next) => {
   const { params: { id }, user: { _id } } = req;
 
   Card.findById(id)
-    .orFail(() => {
-      const error = new Error('Карточка по заданному id отсутствует в базе');
-      error.statusCode = NOT_FOUND;
-      throw error;
-    })
     .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Нет карточки с таким id');
+      }
       if (JSON.stringify(card.owner) !== JSON.stringify(_id)) {
-        res.status(403).send({ message: 'Ошибка: Вы можете удалять только свои карточки' });
-        return;
+        throw new ForbiddenError('Вы можете удалять только свои карточки');
       }
       Card.findByIdAndRemove(card.id)
         .then((myCard) => res.send({ data: myCard }));
     })
-    // .then(() => {
-
-    // })
     .catch((err) => {
       if (err.name === 'CastError') {
-        sendInvalidIdMessage(res);
-        return;
+        next(new IncorrectDataError('Невалидный id карточки'));
       }
-      if (err.statusCode === NOT_FOUND) {
-        sendCardNotFoundByIdMessage(res);
-        return;
-      }
-      sendServerErrorMessage(res);
+      next(err);
     });
-// .orFail(() => {
-//   const error = new Error('Карточка по заданному id отсутствует в базе');
-//   error.statusCode = NOT_FOUND;
-//   throw error;
-// })
-// .then((card) => res.send({ data: card }))
-// .catch((err) => {
-//   if (err.name === 'CastError') {
-//     sendInvalidIdMessage(res);
-//     return;
-//   }
-//   if (err.statusCode === NOT_FOUND) {
-//     sendCardNotFoundByIdMessage(res);
-//     return;
-//   }
-//   sendServerErrorMessage(res);
-// });
 };
 
-module.exports.likeCard = (req, res) => {
+module.exports.likeCard = (req, res, next) => {
   const { user: { _id }, params: { id } } = req;
 
   Card.findByIdAndUpdate(
@@ -86,55 +53,43 @@ module.exports.likeCard = (req, res) => {
     { $addToSet: { likes: _id } }, // добавить _id в массив, если его там нет
     { new: true }, // обработчик then получит на вход обновлённую запись
   )
-    .orFail(() => {
-      const error = new Error('Карточка по заданному id отсутствует в базе');
-      error.statusCode = NOT_FOUND;
-      throw error;
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Нет карточки с таким id');
+      }
+      res.send({ data: card });
     })
-    .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(INCORRECT_DATA).send({ message: 'Ошибка! Переданы некорректные данные для постановки лайка' });
-        return;
+        next(new IncorrectDataError('Ошибка! Переданы некорректные данные для постановки лайка'));
       }
       if (err.name === 'CastError') {
-        sendInvalidIdMessage(res);
-        return;
+        next(new IncorrectDataError('Невалидный id карточки'));
       }
-      if (err.statusCode === NOT_FOUND) {
-        sendCardNotFoundByIdMessage(res);
-        return;
-      }
-      sendServerErrorMessage(res);
+      next(err);
     });
 };
 
-module.exports.dislikeCard = (req, res) => {
+module.exports.dislikeCard = (req, res, next) => {
   const { user: { _id }, params: { id } } = req;
   Card.findByIdAndUpdate(
     id,
     { $pull: { likes: _id } }, // убрать _id из массива
     { new: true },
   )
-    .orFail(() => {
-      const error = new Error('Карточка по заданному id отсутствует в базе');
-      error.statusCode = NOT_FOUND;
-      throw error;
+    .then((card) => {
+      if (!card) {
+        throw new NotFoundError('Нет карточки с таким id');
+      }
+      res.send({ data: card });
     })
-    .then((card) => res.send({ data: card }))
     .catch((err) => {
       if (err.name === 'ValidationError') {
-        res.status(INCORRECT_DATA).send({ message: 'Ошибка! Переданы некорректные данные для снятия лайка' });
-        return;
+        next(new IncorrectDataError('Ошибка! Переданы некорректные данные для снятия лайка'));
       }
       if (err.name === 'CastError') {
-        sendInvalidIdMessage(res);
-        return;
+        next(new IncorrectDataError('Невалидный id карточки'));
       }
-      if (err.statusCode === NOT_FOUND) {
-        sendCardNotFoundByIdMessage(res);
-        return;
-      }
-      sendServerErrorMessage(res);
+      next(err);
     });
 };
